Refactor package structure
Move numpool components from internal to root package and consolidate
database setup functions. This simplifies the package hierarchy and
makes the core functionality directly accessible.
