--- internal/numpool/main_test.go
@@ -1,13 +1,0 @@
-package numpool
-
-import (
-	"os"
-	"testing"
-
-	"github.com/yuku/numpool/internal"
-)
-
-func TestMain(m *testing.M) {
-	internal.SetupTestDatabase()
-	os.Exit(m.Run())
-}
